backend/pyproject.toml
[tool.poetry]
name = "menzar"
version = "0.1.0"
description = ""
authors = ["zvaktualno <tadej.zivic@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "*"
tqdm = "*"
requests = "*"
bs4 = "*"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"
django-crontab = "^0.7.1"
djangorestframework = "^3.14.0"
djangorestframework-jwt = "^1.11.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "*"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
